app/models.py
# Re-export models from centralized CreateDB, adding repo root to sys.path for local runs
import os, sys
_current_dir = os.path.dirname(os.path.abspath(__file__))
_repo_root = os.path.abspath(os.path.join(_current_dir, '..', '..'))
if _repo_root not in sys.path:
    sys.path.insert(0, _repo_root)
from CreateDB.models import UserBalance, LedgerTx  # noqa: F401
